+++ src/assets/styles/configs.scss
@@ -3,13 +3,13 @@ $FontBold: 'Biennale Bold';
 $FontLight: 'Biennale Light';
 
 $PrimaryColor: #5E49FF;
-//$PrimaryColorWithOpacity: #cac5ec;
-$PrimaryColorWithOpacity: #cef1f2;
+$PrimaryColorWithOpacity: #cac5ec;
+$PrimaryColorWithOpacity1: #cef1f2;
 $PrimaryColor3: #3BD42D;
 $PrimaryColor4: #25CBD3;
 $PrimaryColor5: #F0F5FF;
 $PrimaryColor6: #000000;
-$PrimaryColor7: #FFFFFF;
+$WhiteColor: #FFFFFF;
 
 
 $GreyColor: #7C7786;
