+++ src/assets/styles/public.scss
@@ -11,10 +11,10 @@
 
     .logo {
         margin-bottom: 20px;
-        height: 30%;
-        max-width: 260px;
+        height: 50%;
+        max-width: 280px;
     }
-
+   
     
     &.register{
         .logo{
@@ -39,12 +39,12 @@
     }
 
     form {
-        width: 100%;
+        //width: 100%;
         display: flex;
         flex-direction: column;
         align-items: center;
         justify-content: center;
-        margin-top: 2.650rem;
+        margin-top: 2.650rem;//
 
         .input {
             width: 100%;
@@ -54,7 +54,7 @@
             padding-top: 10px;
             padding-bottom: 5px;
             margin-bottom: 15px;
-            border-bottom: 1px solid $GreyColor2;
+            border-bottom: 1px solid $GreyColor2;//
             border-top: 0px;
             border-left: 0px;
             border-right: 0px;
@@ -81,12 +81,12 @@
         margin-top: 24px;
 
         p{
-            color: $GreyColor3;
+            color: $GreyColor3;//
             margin: 0px 0px 2px;
         }
 
         a{
-            color: $PrimaryColor6;
+            color: $PrimaryColor;
             font-family: $FontBold;
             text-decoration: underline;
             cursor: pointer;
@@ -94,16 +94,17 @@
     }
 
     @media screen and (min-width: $DesktopBreakpoint) {
-        width: 100%;
-        height: 100%;
-        background-color: $PrimaryColor7;
-        background-image: none;
-
+        width: 80%;
+        height: 90%;    
+        //background-color: $WhiteColor;
+        //background-image: none;
         display: flex;
         flex-direction: row;
         justify-content:space-around;
-        align-items: center;
-
+        align-items: center; 
+        
+    
+      
         &.register{
             .logo{
                 display: flex;
@@ -111,12 +112,14 @@
         }
 
         img.logo{
-            height: 80%;
+            width: 500px;
+            height: 100%;
             margin: 0px;
-            max-width: 360px;
+            max-width: 400px;
         }
 
-        p{
+        p{ 
+            
             &.error{
                 max-width: none;
             }
@@ -131,18 +134,41 @@
             border: 0.1em solid $PrimaryColor;
             border-radius: 32px;
             padding: 60px 40px;
+            margin-top: 2.650rem;//
+            margin-bottom: 2.650rem;//
 
-            .input{
-                    border-bottom-width: 1.5px;
+            input{
+                    border-bottom-width: 1.5px;  
+                    color: $WhiteColor;
+
+                    &::placeholder {
+                        color: $WhiteColor;
+                        opacity: 1;
+                    }
+                  
+                            
+                   
+                   
 
                     img{
                         margin-right: 15px;
                     }
             }
 
+            .link{
+                      
+                p{
+                    color: $WhiteColor;//
+                 
+                }
+
+            
+
             button{
-                height: 64px;
+                height: 50px;
             }
         }
     }
+    }
 }
+
