+++ src/assets/styles/register.scss
@@ -1,0 +1,44 @@
+@media screen and (min-width: $DesktopBreakpoint) {
+.ContainerInicial{
+        width: 85%;//
+        height: 85%;//
+        box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px,
+         rgba(0, 0, 0, 0.12) 0px -12px 30px, 
+         rgba(0, 0, 0, 0.12) 0px 4px 6px, 
+         rgba(0, 0, 0, 0.17) 0px 12px 13px, 
+         rgba(0, 0, 0, 0.09) 0px -3px 5px;
+        background:rgba(0,0,0,0.55);
+        background-image: none;
+        display: flex;
+        flex-direction: row;
+        justify-content:space-around;
+        align-items: center;  
+        border-radius: 1%;      
+         }
+
+         form.formInicial{
+                width: 30%;
+                height: 50%;
+                border: 0.1em solid $PrimaryColor;
+                border-radius: 32px;
+                padding: 20px 40px;    
+
+         }
+
+        input{ 
+                width: 30%;
+                background-color:  transparent;        
+
+
+        }
+
+        form.formInicial.avatar {
+                border: 2px solid $PrimaryColor;//
+        }
+
+      
+      
+
+
+}
+
