+++ src/components/General/AvatarInput.tsx
@@ -2,13 +2,12 @@ import imgAvatar from  '../../assets/imagens/imgAvatar.svg'
 
 export const AvatarInput = () => {
    return(
-      <div className="container-upload-image">
+      <div className="container-upload-image" >
         <div className="avatar">
            <img src={imgAvatar} />
         </div>
 
       </div>
    )
-
-
 }
+
