+++ src/views/Login.tsx
@@ -13,6 +13,7 @@ export const Login = () => {
    return(
     
       <div className="ContainerPublic">
+        <div className="ContainerInicial register $DesktopBreakpoint">
        <img src={logo} alt='Logo Devaflix' className="logo" />
         <form>
            
@@ -30,6 +31,7 @@ export const Login = () => {
                     <Link to='/register'> Faça seu cadastro agora!</Link>
                 </div>
         </form>    
+     </div>
      </div>
   
    )
